requirements.txt
praw
openai
streamlit
python-dotenv
spacy